Rename C++ test source files from .c to .cc Make revalloc.h common to all C++ stl tests Add C++ tests to test suite
