+++ test/cxx/CMakeLists.txt
@@ -1,0 +1,2 @@
+add_rev_test(SIMPLE_STRUCT simple_struct 30 "all;rv64;memh")
+add_subdirectory(stl)
