+++ test/cxx/stl/CMakeLists.txt
@@ -1,0 +1,4 @@
+add_rev_test(MAP map 30 "all;rv64;memh")
+add_rev_test(VECTOR vector 30 "all;rv64;memh")
+add_rev_test(VECTOR_OBJ vector_obj 30 "all;rv64;memh")
+
