+++ test/cxx/stl/vector/vector.cc
@@ -1,5 +1,5 @@
 #include "rev-macros.h"
-#include "revalloc.hpp"
+#include "revalloc.h"
 #include <vector>
 
 #define assert( x )      \
