+++ test/cxx/stl/vector_obj/vector_obj.cc
@@ -1,5 +1,5 @@
 #include "rev-macros.h"
-#include "revalloc.hpp"
+#include "revalloc.h"
 #include <vector>
 
 #define assert( x )      \
